REST API and documentation updates
Making the required changes in order to enable Swagger-UI to run on the
local machine, without going out to the Swagger.io server.

(1) Remove "\n" characters inside the descriptions in rest_api.json. It
was preventing proper rendering of the document inside of Swagger-UI
interface.
(2) Moved the request parameters inside the request specification, as
this was a known Swagger-UI bug.
(3) Enable CORS on the REST API server to enable Swagger functionality.
(4) Properly check the error returned from jsonpb.Unmarshall, as it
would return " characters inside the error string, making the returned
JSON message invalid.
(5) Stub the REST API for chaincode invocation.

Signed-off-by: Anna D Derbakova <adderbak@us.ibm.com>