Check for dev mode during deploy on REST
Add a check for development mode on chaincode deployment for REST. This
is necessary, as when the peer is running in development mode, the
chaincode is identified by "name", while in production mode, the
chaincode is identified by "path". Therefore, different things need to
be checked for being blank in these two cases.

Signed-off-by: Anna D Derbakova <adderbak@us.ibm.com>