obcbatch: use locks, use own message types
ObcBatch used the pbft messages, both for forwarding obc transactions to
the primary, and for pbft consensus requests to the replicas.  This
could lead to a request bouncing around between primary and backups, and
also could lead to batch requests being submitted with incorrect
content.

By wrapping pbft messages, we can separate messages of ObcBatch and
pbft.

Signed-off-by: Simon Schubert <sis@zurich.ibm.com>