Normalize the processing of logging options for obc-peer
This is a proposed change that addresses #569 and #310, and also partially
addresses #578. The handling of logging levels, and ways of managing logging
levels are currently inconsistent and incomplete in the obc-peer
application. In this change:

Logging options are now top-level options in the openchain.yaml file and on
the command line. A single command-line option '--logging-level' applies to
all commands. In the openchain.yaml however, defaults can be set for each of
the obc-peer commands individually. This made sense to me since it seems
reasonable that logging requirements are likely different for server
vs. client commands.

The syntax for specifying logging levels allows the specification of both a
default level for all modules as well as module-specific levels.

The file openchain/logging.go had originally been copied from the examples of
the op/go-logging package, and was therefore mostly dead and irrelevant
code. This file has been completely replaced. This file now defines a 'hook'
routine that is called during the Cobra processing of the obc-peer commands to
set up logging options.  This hook correctly allows command-line options to
override config file options (cf. #578).

main.go was simplified by removing all of the module-specific setup, and the
logging command-line options were promoted to global options. The new hook is
installed in the Cobra Command structures.

The default logging level for the 'peer' server is now INFO; For the other
client commands the default level is WARNING. Developers and others can easily
override these setting in their openchain.yaml files or on the command line or
environment of each invocation of obc-peer.