Modifying the /chaincode request payload
Modifying the required payload for /chaincode requests to be a
ChaincodeSpec for all of deploy/invoke/query. There is no way to make
it work nicely with Swagger if the payload varies between deploy and
invoke/query. The Swagger UI will not allow for building a variable
payload for testing.

Signed-off-by: Anna D Derbakova <adderbak@us.ibm.com>