DEVENV: Remove explict dependence on host GOPATH
Now that the devenv and peer are combined, we can access the fabric code
via relative path to the Vagrantfile.  While it is noted that many
users will likely want to have a properly formulated GOPATH on
both the host and guest, it is no longer strictly required that they
do so.  By eliminating the requirement to establish a specific
path/env setup, we simplify deployment which is likely to make
things easier for new users who are likely to attempt to
simply clone the repository.

Signed-off-by: Greg Haskins <ghaskins@lseg.com>