create a hook to allow mounting external directories
This creates an environment hook to allow the user to mount an
external directory.  In the OBC workflow, the user could mount
an arbitrary directory into /openblockchain, however this is now
different as devenv has been merged with the fabric tree.  This
commit allows the user to add an external directory which
could have extra dev tools or test chaincode.

Signed-off-by: Joseph C Wang <joequant@gmail.com>