Invert our platforms/car unit-test exclusion logic
We previously made the test exclude TravisCI.  However, there are
other environments besides travis that do not have the prerequisites
installed and thus fail this test as well.  Therefore, we invert the
logic so that the test _only_ runs under vagrant and is disabled
everywhere else.  When the dependency issue is resolved, we can
remove this exclusion.

Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>