peer: Fix possible nil dereference in PeerImpl Unicast
If the receiveHandle is not in the handlerMap, we'll try to call
SendMessage on a nil, which will panic.  I'm not sure there's a way to
actually trigger this bug but it seems cleaner.

The new extracted getMessageHandler function also takes care of lock and
unlocking with defer, which is safer than the previous code (against
possible future changes).