Build docker baseimage on demand
There may be occasions where certain platforms lag behind others w.r.t.
publishing docker baseimages to dockerhub.  Today, resolution of this
condition is at best a manual process.  Now that we have a formal build
system in place, lets make this a little more intelligent.

What this patch does is it treats dockerhub as an optimization, a form
of build-caching.  However, if for whatever reason, an artifact is not
found on dockerhub, we will attempt to build the base image locally.

This is, of course, slower, but it is probably more graceful than
failing the build outright. 

Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>