Build peer and membersvcs docker images natively
We currently use a unit-test written in golang to generate
our peer and membersrvc images.  This is a vestige from the
time before we had a formal build system.  There are distinct
advantages to allowing the build system to fully manage the
generation of these images so we may add better build-avoidance
and caching logic, which Make is naturally good at.  Therefore,
we extract the logic from the golang code base and move it to the
Makefile.  We will further optimize the generation later in the
series.

Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>