Add possibility to use custom hash instead of UUID as TxID
This commit adds a -tid option to peer and a field to Chaincode
invocations (and replaces any previous, similar capabilities
with those) where a UUID generation algorithm can be selected.
Such algorithm contains a hash function plus a decoding. For
now, sha256 with base64 is supported. We take the first 128 bits
of the hash, and use it as a UUID.