Make WriteGopathSrc() more precise
WriteGopathSrc() blindly copies everything under $GOPATH/src.  This
is problematic if any other (particularly large) artifacts make their
way into the tree.  As we introduce the fabric.git/build area, this
will get worse.

We observe that all we really need in this situation is the .go files,
so let's target those specifically.

Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>