CC Tables: Fix buildKeyString() to pay attention to Key cols.
When DeleteRow() calls buildKeyString(), that function does not pay
attention to the Key field in the column definition.  Thus, the key
created includes non-Key column values.  Then DeleteRow() tries to
delete this nonexistent row and silently fails.  You think the row is
deleted, but in reality it is not.  This patch fixes buildKeyString()
to be the same as getKeyAndVerifyRow() which is used by InsertRow()
and does pay attention to the Key field in the column definition when
inserting or updating a row.

Signed-off-by: Alex Tsariounov <tsariounov@gmail.com>