Updating UTs
Updating the chain-tests.js UTs to have a standardized
testChaincodePath location, so that the user running the tests will not
have to update the UTs. They will just have to place their project
files into a specific location. The README.md will be updated to match
the new location.

Updating the asset-mgmt.js test case to have it deploy properly in
network mode. The deploy operation succeeds, but the last test still
fails due to the chaincode hash being passed up incorrectly. Therefore,
the query operation gets the wrong hash and fails. That will be fixed
shortly.

Signed-off-by: Anna D Derbakova <adderbak@us.ibm.com>