Busywork: Changes to support latest blockchain semantics
The definition of correctness for **busywork** tests has been relaxed due to
the realities of the current PBFT consensus algorithms. Due to state transfer,
we can no longer assume that each peer has actually executed each
transaction. Instead, we can only demand that the state of the peer is
consistent with having executed every transaction. We also now wait until the
peers reach a consensus on the blockchain height at the end of the run before
querying for final status, since queries now fail on peers that are in the
middle of state transfer. Once a couple of outstanding consensus PRs are
processed we will probably be able to state that the busywork/counters test
case *stress2b* is working.