busywork changes 6/28/2016
This change set adds several new features and fixes a few minor bugs. As many
of the features are interlocked it is easier for me to present this as one
large change.

One of the motivations was issue #2020, which requested support for being able
to run busywork client drivers against remote machines. The `-interface`
option was added to `userModeNetwork`, and the `-remote` option was added to
the `busywork/counters` driver to enable this. The `busywork/counters`
chaincode was also enhanced with a "ping" query, which is used by all clients
now to determine when the chaincode is up and running. This required adding a
new silent failure mode to the Tcl `fabric::devops` procedure.

The `$BUSYWORK_HOME/chaincodes` file has been changed from a line-oriented
file to a single JSON object, similar to the `network` file, and support has
been added for manipulating this file from client drivers and tools.

The `busywork/bin/busy` script was added. `busy` is a Swiss Army Knife tool
that supports get/deploy/invoke/query/ping on **busywork** networks, including
definitive interlock for deploy/invoke/ping.

An option was added to the `fabricLogger` to log only parts of the blockchain.
This is necessary for efficient implmentation of the `busy` interlocks.

Following the mainline code, all references and usages of PBFT "classic" and
"sieve" have been removed.

The code has been tested during devlopment. I do not have a canned test for
multi-system deployments; I have simply tested this interactively. I have also
added a simple dedicated test script for the `busy` tool that exercises all of
its features.