Interface changes, peer-peer and ranged query support.
RangedQuery needs to be its own interface so that I can mock it out. The mocked ranged query works by keeping a linked list of state names, so that it can walk the map of values.

Also started work on letting chaincodes invoke other chaincodes and supporting transactions. This kind of emulates the original chaincode and handler objects.